refactor: move interfaces to dependants' package
